requirements.txt
aiohttp>=3.8.5,<4.0.0
click>=8.1.3,<9.0.0
python-dotenv>=1.0.0
